+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Now you can define if a virtual foreign key must ignore null values or not
 - Fixed bug that added two ? in Mvc\Url::get when using query parameters (#10421)
 - Now string attributes in models can be marked to allow empty string values (#440)
+- Added an option to return the SQL to be generated from a Mvc\Model\Query instance (#1908)
 
 # 2.0.2 (2015-05-26)
  - Added `stats()` methods to Beanstalk